pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sheet-call-tree"
version = "0.1.0"
description = "Visualize Excel formula dependencies as YAML AST"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "roksechs"}]
keywords = ["excel", "xlsx", "formula", "ast", "yaml", "dependency"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Office/Business :: Financial :: Spreadsheet",
]
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[project.scripts]
sheet-call-tree = "sheet_call_tree.cli:main"

[project.urls]
Homepage = "https://github.com/roksechs/sheet-call-tree"
Repository = "https://github.com/roksechs/sheet-call-tree"
Issues = "https://github.com/roksechs/sheet-call-tree/issues"

[tool.setuptools.packages.find]
where = ["src"]
